refactor: improve app domain
