+++ lib/core/app/domain/models/env.dart
@@ -1,6 +1,4 @@
 // env.dart
-import 'dart:io';
-
 const Map<String, String> environment = <String, String>{
   'baseUrl': 'https://api.themoviedb.org',
   'apiKey': String.fromEnvironment('API_KEY'),
