+++ lib/core/app/domain/models/flavor_config.dart
