+++ lib/ui/states/migration_state.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:tv_randshow/config/locator.dart';
+import 'package:tv_randshow/core/app/ioc/locator.dart';
 import 'package:tv_randshow/core/migration/domain/models/migration_status.dart';
 import 'package:tv_randshow/core/migration/domain/use_cases/add_streamings_migration_use_case.dart';
 import 'package:tv_randshow/core/migration/domain/use_cases/mobile_database_migration_use_case.dart';
