+++ lib/ui/viewmodels/views/info_view_model.dart
@@ -1,6 +1,6 @@
 import 'package:stacked/stacked.dart';
-import 'package:tv_randshow/config/locator.dart';
 import 'package:tv_randshow/core/app/domain/interfaces/i_app_service.dart';
+import 'package:tv_randshow/core/app/ioc/locator.dart';
 import 'package:tv_randshow/core/io/domain/use_cases/export_tvshows_use_case.dart';
 
 class InfoViewModel extends BaseViewModel {
