+++ lib/ui/viewmodels/widgets/delete_model.dart
@@ -1,7 +1,6 @@
 import 'package:stacked/stacked.dart';
-
-import 'package:tv_randshow/config/locator.dart';
 import 'package:tv_randshow/core/app/domain/services/favs_service.dart';
+import 'package:tv_randshow/core/app/ioc/locator.dart';
 
 class DeleteModel extends BaseViewModel {
   final FavsService _favsService = locator<FavsService>();
