+++ lib/ui/viewmodels/widgets/random_model.dart
@@ -1,6 +1,6 @@
 import 'package:stacked/stacked.dart';
 
-import 'package:tv_randshow/config/locator.dart';
+import 'package:tv_randshow/core/app/ioc/locator.dart';
 import 'package:tv_randshow/core/tvshow/domain/models/query.dart';
 import 'package:tv_randshow/core/tvshow/domain/models/tvshow_details.dart';
 import 'package:tv_randshow/core/tvshow/domain/use_cases/get_tvshow_details_use_case.dart';
